requirements.txt
# Core dependencies
yfinance>=0.2.66
pandas>=2.2.0
numpy>=2.0.0
requests>=2.32.0
python-dotenv>=1.1.0

# LangChain and LangGraph
langgraph>=1.0.0
langchain>=0.3.27
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# AI/ML Libraries
prophet>=1.1.5
google-generativeai>=0.8.0

# FastAPI
fastapi>=0.119.0
uvicorn[standard]>=0.37.0
pydantic>=2.12.0

# Streamlit
streamlit>=1.50.0
plotly>=6.0.0

# Additional utilities
pytz>=2024.0
